Se realiza la logica del proyecto

+++ domain/model/src/main/java/co/com/powerup2025/model/solicitudes/Solicitudes.java
@@ -30,5 +30,5 @@ public class Solicitudes {
 
         private Long idEstado;
 
-        private Long idTipoPrestamo;
+        private Integer idTipoPrestamo;
 }
