+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/powerup2025/r2dbc/Entity/SolicitudEntity.java
@@ -41,5 +41,5 @@ public class SolicitudEntity {
     private Long idEstado;
 
     @Column("id_tipo_prestamo")
-    private Long idTipoPrestamo;
+    private Integer idTipoPrestamo;
 }
